+++ src/dynode/model_configuration/config_definition.py
@@ -1,7 +1,7 @@
 """Top level classes for DynODE configs."""
 
 from datetime import date
-from typing import Callable, List, Optional, Union
+from typing import List, Optional, Union
 
 from jax import Array
 from jax import numpy as jnp
@@ -10,14 +10,12 @@ from pydantic import (
     BaseModel,
     ConfigDict,
     Field,
-    PositiveFloat,
     PositiveInt,
     field_validator,
     model_validator,
 )
 from typing_extensions import Self
 
-from ..typing import CompartmentGradiants
 from .bins import AgeBin, Bin
 from .dimension import (
     Dimension,
@@ -131,7 +129,7 @@ class Compartment(BaseModel):
 
 
 class Initializer(BaseModel):
-    """Initalize compartment state of an ODE model."""
+    """Initalize compartment state of the system."""
 
     description: str = Field(
         description="""Description of the initializer, its data streams and/or
@@ -168,10 +166,10 @@ class Initializer(BaseModel):
         )
 
 
-class CompartmentalModel(BaseModel):
+class SimulationConfig(BaseModel):
     """An ODE compartment model configuration file."""
 
-    # allow users to pass custom types into CompartmentalModel
+    # allow users to pass custom types into CompartmentalConfig
     model_config = ConfigDict(arbitrary_types_allowed=True)
     initializer: Initializer = Field(
         description="""Initializer to create initial state with."""
@@ -181,13 +179,6 @@ class CompartmentalModel(BaseModel):
     )
     parameters: Params = Field(
         description="""Model parameters, includes epidemiological and miscellaneous."""
-    )
-    # passed to diffrax.diffeqsolve
-    ode_function: Callable[
-        [List[Compartment], PositiveFloat, Params], CompartmentGradiants
-    ] = Field(
-        description="""Callable to calculate instantaneous rate of change of
-        each compartment."""
     )
 
     @model_validator(mode="after")
@@ -236,6 +227,41 @@ class CompartmentalModel(BaseModel):
         return self
 
     @model_validator(mode="after")
+    def _create_introduction_ages_mask_encoding(self) -> Self:
+        """Parse Strain's introduction_ages to a binary mask."""
+        # dont bother one-hot encoding introduction_ages if they dont exist
+        if any(
+            [
+                strain.introduction_ages is not None
+                for strain in self.parameters.transmission_params.strains
+            ]
+        ):
+            # find a dimension with Age stratification
+            age_binning = []
+            for dim in self.flatten_dims():
+                # only check first element since dimensions must all be same type
+                if isinstance(dim.bins[0], AgeBin):
+                    age_binning = dim.bins
+                    break
+            assert len(age_binning) > 0, (
+                "attempted to encode introduction_ages but could not "
+                "find any age structure in the compartments"
+            )
+            mask = []
+            for strain in self.parameters.transmission_params.strains:
+                # assume intro_ages is found in age_structure due to above validator
+                if strain.introduction_ages is not None:
+                    mask = [
+                        1 if b in strain.introduction_ages else 0
+                        for b in age_binning
+                    ]
+                else:
+                    mask = [0 for _ in age_binning]
+                # set the private field now that validation is complete.
+                strain.introduction_ages_mask_vector = mask
+        return self
+
+    @model_validator(mode="after")
     def _validate_introduced_strains(self) -> Self:
         """Validate that all introduced strains have the same age binning as defined by the Model's compartments."""
         strains = self.parameters.transmission_params.strains
@@ -249,14 +275,16 @@ class CompartmentalModel(BaseModel):
                         target_age in age_structure
                         for target_age in strain_target_ages
                     ]
-                ), f"""{strain.strain_name} attempts to introduce itself using
-                    {strain_target_ages} age bins, but those are not found
-                    within the age structure of the model."""
+                ), (
+                    f"{strain.strain_name} attempts to introduce itself using "
+                    f"{strain_target_ages} age bins, but those are not found "
+                    "within the age structure of the model."
+                )
         return self
 
     @model_validator(mode="after")
     def _validate_vaccination_counts(self) -> Self:
-        """Validate vaccination dose definitions are correct across the model."""
+        """Validate vaccination dose definitions across the Compartments."""
         # assert that all similarly named dimensions have same vaccine bins
         num_shots: dict[str, int] = {}
         all_dims = self.flatten_dims()
@@ -273,7 +301,7 @@ class CompartmentalModel(BaseModel):
         return self
 
     def get_compartment(self, compartment_name: str) -> Compartment:
-        """Search the CompartmentModel and return a specific Compartment if it exists.
+        """Search the CompartmentalConfig and return a Compartment if it exists.
 
         Parameters
         ----------
@@ -345,8 +373,8 @@ class InferenceProcess(BaseModel):
     """Inference process for fitting a CompartmentalModel to data."""
 
     model_config = ConfigDict(arbitrary_types_allowed=True)
-    model: CompartmentalModel = Field(
-        description="CompartmentalModel on which inference is performed."
+    simulation_config: SimulationConfig = Field(
+        description="SimulationConfig on which inference is performed."
     )
     # includes observation method, specified at runtime.
     inference_method: Optional[MCMC | SVI] = Field(
@@ -355,6 +383,6 @@ class InferenceProcess(BaseModel):
         currently only MCMC and SVI supported""",
     )
     inference_parameters: InferenceParams = Field(
-        description="""inference related parameters, not to be confused with
+        description="""Inference related parameters, not to be confused with
         CompartmentalModel parameters for solving ODEs."""
     )