+++ src/dynode/model_configuration/params.py
@@ -2,13 +2,14 @@
 
 from typing import List
 
-import chex
+from diffrax import AbstractSolver, Tsit5
 from jax import Array
 from jax.random import PRNGKey
 from numpyro.distributions import Distribution
 from pydantic import (
     BaseModel,
     ConfigDict,
+    Field,
     NonNegativeFloat,
     PositiveFloat,
     PositiveInt,
@@ -25,21 +26,46 @@ class SolverParams(BaseModel):
     """Parameters used by the ODE solver."""
 
     model_config = ConfigDict(arbitrary_types_allowed=True)
-    ode_solver_rel_tolerance: PositiveFloat
-    ode_solver_abs_tolerance: PositiveFloat
-
-
-@chex.dataclass
-class ODEParameters:
-    """The internal representation containing parameters passed to the ODEs.
-
-    Because ODEs work with vectors/matricies/tensors as opposed to objects,
-    this internal state flattens the list of strains into the tensors of information
-    separate from the `Strain` class entirely.
-    """
-
-    strain_interactions: chex.ArrayDevice
-    betas: chex.ArrayDevice
+    solver_method: AbstractSolver = Field(
+        default_factory=lambda: Tsit5(),
+        description="""What sort of differential equation solver you wish to
+        use to solve ODEs, defaults to Tsit5(), a general solver good for
+        non-stiff problems. For more information on picking a solver see:
+        https://docs.kidger.site/diffrax/usage/how-to-choose-a-solver/""",
+    )
+    ode_solver_rel_tolerance: PositiveFloat = Field(
+        default=1e-5,
+        description="""Solver relative tolerance, used for adaptive step sizer
+        to decide the size of a subsequent step. Use constant_step_size to
+        switch to constant solver mode. For more information on tolerance see
+        the `choosing tolerances` drop down here:
+        https://docs.kidger.site/diffrax/api/stepsize_controller/#diffrax.PIDController""",
+    )
+    ode_solver_abs_tolerance: PositiveFloat = Field(
+        default=1e-6,
+        description="""Solver absolute tolerance, used for adaptive step sizer
+        to decide the size of a subsequent step. Use constant_step_size to
+        switch to constant solver mode. For more information on tolerance see
+        the `choosing tolerances` drop down here:
+        https://docs.kidger.site/diffrax/api/stepsize_controller/#diffrax.PIDController""",
+    )
+    max_steps: PositiveInt = Field(
+        default=int(1e6),
+        description="""The maximum number of steps the ode solver will take
+        before raising an error. For complex problems use higher number.""",
+    )
+    constant_step_size: NonNegativeFloat = Field(
+        default=0,
+        description="""If non-zero, solver will use constant step size
+        equal to the value set. If 0 solver will use adaptive step size with
+        ode_solver_rel/abs_tolerance""",
+    )
+    discontinuity_points: list[int] = Field(
+        default_factory=lambda: [],
+        description="""Points in the ode's solve that a discontinuity occurs,
+        meaning the higher order gradiants are not smooth. Int values
+        represent the simulation day, or days since init date of the model.""",
+    )
 
 
 class TransmissionParams(BaseModel):
@@ -57,18 +83,20 @@ class TransmissionParams(BaseModel):
         strain_names = [strain.strain_name for strain in self.strains]
         for infecting_strain in strain_names:
             for recovered_from_strain in strain_names:
-                assert (
-                    infecting_strain in self.strain_interactions.keys()
-                ), f"""{infecting_strain} not found in first level of the
-                    strain_interactions_dictionary, , every strain should
-                    have an interaction value against all other strains, found :
-                    {list(self.strain_interactions.keys())}"""
+                assert infecting_strain in self.strain_interactions.keys(), (
+                    f"{infecting_strain} not found in first level of the "
+                    f"strain_interactions_dictionary, , every strain should "
+                    f"have an interaction value against all other strains, found : "
+                    f"{list(self.strain_interactions.keys())}"
+                )
                 assert (
                     recovered_from_strain
                     in self.strain_interactions[infecting_strain]
-                ), f"""unable to find {recovered_from_strain} within
-                    strain_interactions[{infecting_strain}], every strain should
-                    have an interaction value against all other strains."""
+                ), (
+                    f"unable to find {recovered_from_strain} within "
+                    f"strain_interactions[{infecting_strain}], every strain "
+                    f"should have an interaction value against all other strains."
+                )
         return self
 
     @field_validator("strains", mode="after")